Console log parser functional again
