+++ app.go
@@ -20,17 +20,87 @@ type App struct {
 	bdFetcher     *BDFetcher
 	players       *PlayerStates
 	metaInFlight  atomic.Bool
+	uiUpdates     chan any
 	configUpdates chan config.Config
 }
 
 func New(conf config.Config, metaFetcher *MetaFetcher, bdFetcheer *BDFetcher) *App {
 	return &App{
-		console:       tf.NewConsoleLog(),
 		config:        conf,
 		metaFetcher:   metaFetcher,
 		bdFetcher:     bdFetcheer,
+		console:       tf.NewConsoleLog(),
 		players:       NewPlayerStates(),
 		configUpdates: make(chan config.Config),
+		uiUpdates:     make(chan any),
+	}
+}
+
+// Start brings up all the background goroutines and starts the main event processing loop.
+func (app *App) Start(ctx context.Context, done <-chan any) {
+	// Watch the config for writes
+	go config.Notify(ctx, config.DefaultConfigName, app.configUpdates)
+
+	// Handle removing expired players from the active player states
+	go app.players.cleaner(ctx)
+
+	// Start sending player state updates to the UI.
+	go app.stateSyncer(ctx)
+
+	// Load the bot detector lists
+	go app.bdFetcher.Update(ctx)
+
+	// Open the console log and start processing events.
+	if err := app.console.Open(app.config.ConsoleLogPath); err != nil {
+		slog.Warn("Failed to open console file", slog.String("error", err.Error()),
+			slog.String("path", app.config.ConsoleLogPath))
+	}
+
+	go app.logEventUpdater(ctx)
+	go app.upUpdater(ctx)
+
+	dumpTicker := time.NewTicker(time.Duration(app.config.UpdateFreqMs) * time.Millisecond)
+	for {
+		select {
+		case <-dumpTicker.C:
+			if app.metaInFlight.Load() {
+				continue
+			}
+			app.updateMetaProfile(ctx)
+			app.updateBD()
+		case conf := <-app.configUpdates:
+			app.uiUpdates <- conf
+		case <-ctx.Done():
+			return
+		case <-done:
+			return
+		}
+	}
+}
+
+func (app *App) logEventUpdater(ctx context.Context) {
+	eventChan := make(chan tf.LogEvent)
+	app.console.RegisterHandler(tf.EvtAny, eventChan)
+	for {
+		select {
+		case evt := <-eventChan:
+			app.uiUpdates <- evt
+		case <-ctx.Done():
+			return
+		}
+	}
+}
+
+func (app *App) upUpdater(ctx context.Context) {
+	for {
+		select {
+		case msg := <-app.uiUpdates:
+			if app.ui != nil {
+				app.ui.Send(msg)
+			}
+		case <-ctx.Done():
+			return
+		}
 	}
 }
 
@@ -86,39 +156,6 @@ func (app *App) sendPlayerStates() {
 	}
 
 	app.ui.Send(players)
-}
-
-func (app *App) Start(ctx context.Context, done <-chan any) {
-	dumpTicker := time.NewTicker(time.Duration(app.config.UpdateFreqMs) * time.Millisecond)
-
-	go config.Notify(ctx, config.DefaultConfigName, app.configUpdates)
-	go app.players.cleaner(ctx)
-	go app.stateSyncer(ctx)
-	go app.bdFetcher.Update(ctx)
-
-	if app.config.ConsoleLogPath != "" {
-		if err := app.console.Open(app.config.ConsoleLogPath); err != nil {
-			slog.Warn("Failed to open console file", slog.String("error", err.Error()),
-				slog.String("path", app.config.ConsoleLogPath))
-		}
-	}
-
-	for {
-		select {
-		case <-dumpTicker.C:
-			if app.metaInFlight.Load() {
-				continue
-			}
-			app.updateMetaProfile(ctx)
-			app.updateBD()
-		case conf := <-app.configUpdates:
-			app.ui.Send(conf)
-		case <-ctx.Done():
-			return
-		case <-done:
-			return
-		}
-	}
 }
 
 func (app *App) updateBD() {
